backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-multipart
sentence-transformers
faiss-cpu
numpy==1.26.4
requests
plotly
dotenv
streamlit
